Fix bad link ref. when system leveldb exists.
This commit fixes a link-time issue that comes about when an
incompatible system-level leveldb is already installed. The problem is
that leveldb isn't listed as a dependency, so link-time -lleveldbs
aren't preceded by -L{prefix to the leveldb I actually want}/lib, so
when a potentially incompatible system-level leveldb exists, then bad
things happen.

@brobey you might be interested in this.